Create Company, User, JobAd, UserJobAd models to DB #7 (#8)
Changelog
====
- Create `Company`, `User`, `JobAd`, `UserJobAd` models to DB
- `Company`: a model for companies. It contains fields such as `name`,
`country`, `location` for each company.
  - `User`: a model for users. It doesn't contains any fields.
- `JobAd`: a model for job ads. It contains fields such as `position`,
`companyId`, `reward`, `content`, `skills` for each job ad. `companyId`
is a foreign key which references `id` of `Companies` table.
- `UserJobAd`: a model for job ads which users applied for. It contains
fields such as `userId`, `jobAdId` for each application. This is a
bridge table which connects the `Users` and `JobAds` tables. It has a
constraint that no same (userId, jobAdId) can be added repeatedly.